move ResumableHostTrapError into resumable module
